+++ Dependencies/FastNoise/FastNoise.h
@@ -39,7 +39,7 @@
 #ifdef FN_USE_DOUBLES
 typedef double FN_DECIMAL;
 #else
-typedef float FN_DECIMAL;
+typedef double FN_DECIMAL;
 #endif
 
 class FastNoise
